Address CI and OCR feedback for Windows Esc/Ctrl-C (Fixes #333)
Format and clippy cleanups, extract input tests under the 1000-line hard
limit, and cover probe_attached_terminal Attached when an agent is present.
