+++ src/runtime/manager.rs
@@ -346,6 +346,15 @@ impl TmuxRuntimeManager {
         self.clipboard_enforced.contains(session)
     }
 
+    /// Test-only: mark an agent as attached without spawning a viewer.
+    ///
+    /// Used by binary-crate key-routing tests (issue #333) that cannot reach
+    /// `pub(crate)` fields. Not for production attach/detach paths.
+    #[doc(hidden)]
+    pub fn set_attached_agent_id_for_test(&mut self, agent_id: Option<AgentId>) {
+        self.attached_agent_id = agent_id;
+    }
+
     /// Test-only setter for recording clipboard passthrough without invoking tmux.
     #[cfg(test)]
     fn record_clipboard_passthrough(&mut self, session: &str) {
